+++ eslint.config.js
@@ -1,9 +1,13 @@
 import vladpuz from 'eslint-config-vladpuz'
 
-const config = vladpuz()
-
-config.push({
-  ignores: ['build'],
-})
-
-export default config
+export default [
+  ...vladpuz(),
+  {
+    name: 'extensions',
+    rules: {
+      'no-restricted-imports': ['error', {
+        patterns: ['**/*.js'],
+      }],
+    },
+  },
+]