+++ src/getQueueOptions.ts
@@ -1,24 +1,18 @@
-import type { AxiosRequestConfig } from 'axios'
+import type { AxiosInstance, AxiosRequestConfig } from 'axios'
 import type { QueueAddOptions } from 'p-queue'
 
 export function getQueueOptions(
   priority: number,
-  config?: AxiosRequestConfig,
-): QueueAddOptions & { throwOnTimeout: true } {
-  const defaultQueueOptions: QueueAddOptions & { throwOnTimeout: true } = {
-    throwOnTimeout: true,
-    priority,
-  }
-
+  axios: AxiosInstance,
+  config: AxiosRequestConfig = {},
+): QueueAddOptions {
   // eslint-disable-next-line @typescript-eslint/no-unsafe-type-assertion
-  const signal = config?.signal as AbortSignal | undefined
-
-  if (signal == null) {
-    return defaultQueueOptions
-  }
+  const signal = (
+    config.signal ?? axios.defaults.signal
+  ) as AbortSignal | undefined
 
   return {
-    ...defaultQueueOptions,
+    priority,
     signal,
   }
 }