+++ src/resources/cancelOrder.ts
@@ -1,12 +1,8 @@
-import type { ResponseData } from '../types.js'
-
-/* RequestData */
+import type { ResponseData } from '../types.ts'
 
 export interface CancelOrderRequestData {
   id: number
   issue?: string
 }
-
-/* ResponseData */
 
 export type CancelOrderResponseData = ResponseData