+++ src/resources/serverConfig.ts
@@ -1,21 +1,17 @@
-import type { Pricing, ResponseData, Specs, Visibility } from '../types.js'
-
-/* RequestData */
+import type { AutoPrice, Currency, Pricing, PricingInOriginalUsd, ResponseData, Specs, Visibility } from '../types.ts'
 
 export interface ServerConfigRequestData {
   server_name: string
 }
 
-/* ResponseData */
-
-export interface ServerConfigResponseDataConfigBackgroundJob {
+export interface ServerConfigBackgroundJob {
   times_updated: number
   image: string
   command: string
   env: Record<string, string>
 }
 
-export interface ServerConfigResponseDataConfig {
+export interface ServerConfig {
   name: string
   connected: boolean
   visibility: Visibility
@@ -27,10 +23,13 @@ export interface ServerConfigResponseDataConfig {
   id: number
   rental_status: number
   specs: Specs
-  background_job: ServerConfigResponseDataConfigBackgroundJob
+  background_job: ServerConfigBackgroundJob
+  allowed_coins: Currency[]
+  autoprice: AutoPrice
+  usd_pricing: PricingInOriginalUsd
 }
 
 export interface ServerConfigResponseData extends ResponseData {
-  config: ServerConfigResponseDataConfig
+  config: ServerConfig
   creation_completed: boolean
 }